+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## [JSimpleShell x.x - XXXX-XX-XX][x.x]
 [x.x]: https://oss.sonatype.org/content/groups/public/de/raysha/lib/jsimpleshell/x.x/
 
+## [JSimpleShell 3.0.1 - 2014-10-20][3.0.1]
+[3.0.1]: https://oss.sonatype.org/content/groups/public/de/raysha/lib/jsimpleshell/3.0.1/
+
+* [IMPORTANT/fix] The dependencies in all handlers will set again after a loop. This will fix the issue about the loop-prompt.
 * [fix] Dependency injection works also now for inherited fields/methods.
 * [IMPORTANT/feature] If the shell is in record mode a special variable will be stored into the environment. If you wants to know whether you are currently in the record mode, you can use the _CommandRecorder.isShellInRecordMode()_ -Method.